backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
aiosmtplib==3.0.1
email-validator==2.1.1
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
aiofiles==23.2.1

